+++ BidMachine/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.3.0.0.1
+* Distributed adapter as a static framework within the XCFramework.
+* Updated ad display failed error code.
+
 ## 3.3.0.0.0
 * Certified with BidMachine SDK 3.3.0.